Finish review endpoints

+++ app/api/reviews.py
@@ -9,27 +9,39 @@ bp = Blueprint("reviews", __name__, url_prefix="/reviews")
 
 
 @bp.route("/<int:review_id>", methods=["patch"])
+@login_required
 def update_review(review_id):
     form = ReviewForm()
     review_tobe_updated = Review.query.get(review_id)
 
-    if form.validate_on_submit():
-        review_tobe_updated.rating = form.data["rating"]
-        review_tobe_updated.review = form.data["review"]
-        db.session.commit()
+    if review_tobe_updated and review_tobe_updated.customer_id == current_user.id:
+        if form.validate_on_submit():
+            review_tobe_updated.rating = form.data["rating"]
+            review_tobe_updated.review = form.data["review"]
+            db.session.commit()
+            return {
+                "rating": review_tobe_updated.rating,
+                "review": review_tobe_updated.review
+            }, 201
 
-    if form.errors:
-        return {"errors": form.errors}, 400, {"Content-Type": "application/json"}
-
-    return {"review": review_tobe_updated.to_dict()}, 200, {"Content-Type": "application/json"}
+        if form.errors:
+            return {
+                "message": "Validation Error",
+                "statusCode": 400,
+                "errors": form.errors
+                }, 400, {"Content-Type": "application/json"}
+    return "Fail to update", 404
 
 
 @bp.route("/<int:review_id>", methods=["delete"])
+@login_required
 def delete_review(review_id):
     review_tobe_deleted = Review.query.get(review_id)
-    db.session.delete(review_tobe_deleted)
-    db.session.commit()
-    return {
-        "message": "Successfully deleted",
-        "statusCode": 200
-    }, 200, {"Content-Type": "application/json"}
+    if review_tobe_deleted and review_tobe_deleted.customer_id == current_user.id:
+        db.session.delete(review_tobe_deleted)
+        db.session.commit()
+        return {
+            "message": "Successfully deleted",
+            "statusCode": 200
+            }, 200, {"Content-Type": "application/json"}
+    return "Fail to delete", 404
